database/requirements.txt
pandas==2.1.0
psycopg2-binary==2.9.9
tqdm==4.66.1
